mvp init
