--- LABS/LAB5/MVPTable/Presenter/Lab5TableModuleIO.swift
@@ -1,7 +1,0 @@
-protocol Lab5TableModuleInput: AnyObject {
-
-}
-
-protocol Lab5TableModuleOutput: AnyObject {
-
-}
