--- LABS/LAB5/MVPTable/Presenter/Lab5TablePresenter.swift
@@ -1,9 +1,0 @@
-final class Lab5TablePresenter {
-    weak var view: Lab5TableViewInput!
-    weak var moduleOutput: Lab5TableModuleOutput!
-
-}
-
-extension Lab5TablePresenter: Lab5TableViewOutput {
-
-}
