--- LABS/LAB5/MVPTable/View/Lab5TableViewController.swift
@@ -1,14 +1,0 @@
-import UIKit
-
-final class Lab5TableViewController: UIViewController {
-    
-    var output: Lab5TableViewOutput!
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
-
-extension Lab5TableViewController: Lab5TableViewInput {
-    
-}
