--- LABS/LAB5/MVPTable/View/Lab5TableViewIO.swift
@@ -1,7 +1,0 @@
-protocol Lab5TableViewInput: AnyObject {
-
-}
-
-protocol Lab5TableViewOutput: AnyObject {
-
-}
