+++ LABS/LAB_5/MVPTable/Lab5TableViewBuilder.swift
@@ -1,3 +1,13 @@
+//
+//  Lab5TableBuilder.swift
+//  LABS
+//
+//  Created by Иван Тазенков on 27.03.2022.
+//
+
+import Foundation
+import UIKit
+
 final class Lab5TableBuilder {
     func build(output: Lab5TableModuleOutput) -> UIViewController {
         let viewController = Lab5TableViewController()
@@ -6,7 +16,7 @@ final class Lab5TableBuilder {
         viewController.output = presenter
         presenter.view = viewController
         presenter.moduleOutput = output
-        
+
         return viewController
     }
 }
