+++ LABS/LAB_5/MVPTable/Presenter/Lab5TableModuleIO.swift
@@ -1,0 +1,12 @@
+//
+//  Lab5TableModuleIO.swift
+//  LABS
+//
+//  Created by Иван Тазенков on 27.03.2022.
+//
+
+import Foundation
+
+protocol Lab5TableModuleInput: AnyObject {}
+
+protocol Lab5TableModuleOutput: AnyObject {}
