+++ LABS/LAB_5/MVPTable/View/Lab5TableViewIO.swift
@@ -1,0 +1,12 @@
+//
+//  Lab5TableViewIO.swift
+//  LABS
+//
+//  Created by Иван Тазенков on 27.03.2022.
+//
+
+import Foundation
+
+protocol Lab5TableViewInput: AnyObject {}
+
+protocol Lab5TableViewOutput: AnyObject {}
